backend/requirements.txt

flask
flask-cors
python-dotenv
openai 
google-generativeai
flask-sqlalchemy
flask-bcrypt
flask-jwt-extended
